wip: simple render caching
The current impl for syntax highlighting is far to slow and results in
noticable lag in the UI when scrolling for Rust files due to the
complexity of the regex being used to tokenise each visible line. The
proper fix for this is to use a different approach for syntax
highlighting and there are several UI bugs around invalidating the cache
correctly in this commit, but its going to be useful to keep around for
reference in the short term I think.

Ultimately, I think I'm going to remove syntax highlighting as part of
the core editor and instead have it as part of the UI interface as /
when it gets implemented.